Copied from Perforce
 Change: 190373
 ServerID: perforce.ravenbrook.com

+++ mps/code/global.c
@@ -760,7 +760,7 @@ Bool ArenaStep(Globals globals, double interval, double multiplier)
       trace = ArenaTrace(arena, (TraceId)0);
     } else {
       /* No traces are running: consider collecting the world. */
-      if (PolicyShouldCollectWorld(arena, availableEnd - now, now,
+      if (PolicyShouldCollectWorld(arena, (double)(availableEnd - now), now,
                                    clocks_per_sec))
       {
         Res res;
